refactor: extract logic for serializing credential options
